Fix cancel_all_events event name parsing
The cancel_all_events function was incorrectly assuming
that event tags didn't have dashes in them, but they all
do since they are either port or volume IDs, which are UUIDs.

Instead of just storing our event data indexed by the composite
event key, we should just store the name and tag distinctly so
that there is no question about how to dissect the two from the
key later. This is all internal record-keeping inside of the
compute manager, so we can change it to whatever we want.

This includes a slight change to the internal virtapi interface
for wait_for_instance_event(), which removes the ability to
pass composite event keys as strings instead of (name, tag)
tuples. In reality, none of the virt code was using the key string
format, so this just cleans up a few tests that were. It makes
that interface less ambiguous, which is good anyway.

Change-Id: I15f041e820c2dabe82e8e25bf77f9cc86330a76a
Closes-Bug: #1760303