Move test_report_client out of placement namespace
test_report_client provides functional tests of the report client using
a fully operating placement service (via wsgi-intercept) but it is not,
in itself, testing placement. Therefore this change moves the test
into nova/tests/functional where it can sit besides other genral purpose
nova-related functional tests.

As noted in the moved file, in a future where placement is extracted,
nova could choose to import a fixture that placement (installed as a
test dependency) provides so that this test and ones like it can
continue to run as desired.

compute/test_resource_tracker.py is updated to reflect the new location
of the module as it makes use of it.

partially implements blueprint placement-extract

Change-Id: I433700e833f97c0fec946dafc2cdda9d49e1100b