objects: Add NUMACell.network_metadata
This will be used to keep track of the physical networks and potential
tunneled network that are affined with this given NUMA cell. It's
implemented by way of the recently added 'NetworkMetadata' object. We
don't currently populate this - that will come at the end of the series.

Part of blueprint numa-aware-vswitches

Change-Id: I9140ddae666f9c994c7954bc2e690ac1064a89de