Cleanup _get_request_spec_for_select_destinations for live migrate
We only need to get the image metadata and build the legacy
filter_properties dict if we don't have a request spec, which
at this point in nova should never be the case, but cleaning
all of that fallback code is a longer bigger effort for another
day.

This change simply moves the image meta and filter properties
code into the conditional that actually uses them, and cleans
up several tests while at it.

Change-Id: I68589e00dd6c2efdbe19eb4cea0be546a1bd5151