Log a more useful error when cinder auth isn't configured
This is similar to I18f162c4f8d1964cb4d0c184ff2149c76e1e86b4,
except all REST API interactions that hit the Cinder API use
the request user token, so configuring auth for [cinder] was
only recently added (in Queens) and is not required. However,
it is needed for certain scenarios where there is no user token,
like with periodic tasks to shutdown an instance and detach its
volumes. As such, we need to provide a more useful error message
to help an operator debug the problem if this don't have the
configuration needed for some operations.

Change-Id: I7adaa003a86503b711e87be671e9ef9c0e52b152
Related-Bug: #1761487