Avoid dumping stack on BuildAbortException
While debugging something I noticed that we are dumping stack when we
intentionally abort a build. When we raise BuildAbortException, it means
we have caught some situation and are handling it, setting error states
and such. We do not need to dump stack in that situation, and doing so
pollutes the logs. We already dump stack when we catch an unexpected
exception in the same path, which makes sense.

This just changes that LOG.exception to LOG.error and adds a trivial
test verification to make it clear that it is the behavior we want.

Change-Id: I4e3343f250ee563e87001b0e17ab19dc872caa66