conf: Add '[neutron] physnets' and related options
Things are getting a little bit magic here. We don't want to store
inventory in configuration and we _really_ don't want to do so with
another JSON blob a la the various '[pci]' options. Placement is not yet
at a stage where we can avoid the former but we can certainly avoid the
latter. Do so through adding some static opts and some dynamic filter
configuration, similarly to what Cinder does for backend configuration
using its 'enabled_backends' configuration option.

None of this is actually used yet - that will come later.

Part of blueprint numa-aware-vswitches

Change-Id: Id7c2f0b53c8871ff47a836ec4c324c8cce430b79