Delete allocations from API if nova-compute is down
When performing a "local delete" of an instance, we
need to delete the allocations that the instance has
against any resource providers in Placement.

It should be noted that without this change, restarting
the nova-compute service will delete the allocations
for its compute node (assuming the compute node UUID
is the same as before the instance was deleted). That
is shown in the existing functional test modified here.

The more important reason for this change is that in
order to fix bug 1756179, we need to make sure the
resource provider allocations for a given compute node
are gone by the time the compute service is deleted.

This adds a new functional test and a release note for
the new behavior and need to configure nova-api for
talking to placement, which is idempotent if
not configured thanks to the @safe_connect decorator
used in SchedulerReportClient.

Closes-Bug: #1679750
Related-Bug: #1756179

Conflicts:
      nova/compute/api.py

NOTE(mriedem): The compute/api conflict is due to not
having change I393118861d1f921cc2d71011ddedaf43a2e8dbdf
in Pike. In addition to this, the call to
delete_allocation_for_instance() does not include the
context parameter which was introduced in change
If38e4a6d49910f0aa5016e1bcb61aac2be416fa7 which is
also not in Pike.

Change-Id: If507e23f0b7e5fa417041c3870d77786498f741d
(cherry picked from commit ea9d0af31395fbe1686fa681cd91226ee580796e)
(cherry picked from commit cba1a3e2c1b161204a3662a0d9fbf33da38aa7d3)