Fix member_of with sharing providers
When the `member_of` parameter was present, non-sharing providers
incorrectly associated sharing providers with other sharing providers
with no direct connection to non-sharing providers involved in other
parts of the request.

This patch fixes the bug by adding the aggregate check in SQL
query when getting sharing providers.

Change-Id: I566cc5458175f8e3075be6967e5f6d91847ce198
Closes-Bug: #1763907
Blueprint: alloc-candidates-member-of