Update os_compute_api:os-flavor-extra-specs:index docs for 2.47
The 2.47 compute API relies on policy rule
os_compute_api:os-flavor-extra-specs:index to determine
if flavor extra specs should be shown in the embedded flavor
response when showing server details, updating a server or
rebuilding a server. This change updates the documentation
for the policy rule to indicate its usage in those APIs.

Change-Id: I1116751f2a065edf15ea7d492abde6e01c3e4cdb
Partial-Bug: #1763741