Remove redundant _do_check_can_live_migrate_destination
This was added in change I8c2f1a67 with the note:

===
We'll want to add claims around the logic of this method, so it's better
if there is a nested method that does the business logic to
avoid deep nesting that makes the code less readable.
===

AFAICT this logic was never added, and the indirection has always been
redundant.

Change-Id: Ie01f2c056a85e08a46c2461f1952cfdbaac3a547