libvirt: mdevs returning parent and vendor PCI info
In order to provide inventories per physical GPUs, the libvirt driver needs to
know the parent ID for each mediated device.
Also, given we need to limit the max unit of vGPUs per guest to 1 if that's a
nvidia GPU, we also need to pass the PCI vendor ID down to the driver.

Related-Bug: #1758086
Change-Id: Ib80af55ccf4e578e7f8b58531da1b1e95783c779