xenapi: handle InstanceNotFound in detach_interface()
CI tests frequently teardown by rolling back interface
attachments and then deleting the server, both of which
are asynchronous. If we're trying to detach an interface
on a guest that is gone, we don't need to log a traceback
exception in the logs, just let the InstanceNotFound
raise up to the ComputeManager which will handle the
error and log it appropriately.

Change-Id: I9428be0e6e5b640fdda00410817925001361fd2c
Closes-Bug: #1759979