libvirt: Add a debug log entry before / after invoking migrate()
The guest.migrate() is running in an async background thread, so it is
useful to see exactly when it runs relative to other threads.  And the
existing message: "Starting monitoring of live migration" is a bad one
to rely on because there's no guarantee libvirt has been told to start
migration yet.

So add a debug log entry before and after issuing the guest.migrate()
call.

Change-Id: Ia7d5de5b0ee4f5d88b5101d2258336adf573a350
Suggested-by: Daniel Berrangé (discovered via debugging a Nova live
              migration bug)
Signed-off-by: Kashyap Chamarthy <kchamart@redhat.com>