Fix detach_volume calls when rolling back a failed attach
Change I751fcb7532679905c4279744919c6cce84a11eb4 modified
the ComputeDriver.detach_volume method signature to pass
the RequestContext as the first argument. This was missed
in the volume attach rollback code where the driver.detach_volume
method is called. It was missed because that is called
in a try/except block which catches generic Exception, which
also means that the unit test covering that flow, which
raised an Exception, was not failing.

This change fixes the code bug and re-writes the test to
use mock rather than mox and explicitly assert the calls made.

An alternative to this would be changing the generic
Exception handling in the rollback code, however, the
virt drivers raise things other than NovaException, like
os-brick exceptions and libvirtError, so refactoring that
error handling is non-trivial.

Change-Id: I9bafd4c47489b61973302310720b11ec1b5e0374
Closes-Bug: #1765742