libvirt: fix hard reboot issue with mdevs
Sometimes, a Nova instance may not have an associated libvirt guest,
like if a live migration fails and the hypervisor on the source node
is in an inconsistent state.
The usual way to fix that is to hard reboot the instance, so the
virt driver recreates the guest.
Unfortunately, we introduced a regression in Queens with
Ie6c9108808a461359d717d8a9e9399c8a407bfe9 because we weren't
checking whether the guest was there or not.

This fix aims to catch the InstanceNotFound exception that's
raised and just tell there are no mediated devices attached to that
instance (given we can no longer know them).

Change-Id: Iff30e59ced63a94c5c3377317a2e2a502f96eb91
Closes-Bug: #1764460