Deduplicate config/policy reference docs from main index
The top level index (home page) was duplicating the
configuration/index content, so this simply changes
the home page into a table of contents for the configuration
sub-tree and leaves the config/policy content in the
sub-tree. This will be needed when we add docs about
placement policy.

The hidden configuration toc tree items are moved
into the sub-tree configuration/index to be closer
to the actual documents we're hiding from the toc tree.

Related to blueprint granular-placement-policy

Change-Id: Iad87dc339278ee7e7cf8de5eea252bbb7a5f75c2