log stale allocations as WARNING instead of DEBUG
Stale resource allocations can prevent Nova from scheduling new
instances on a host that otherwise has sufficient resources available.
Nova would log a message about this situation on the affected compute
node, but it was logged at DEBUG priority and so would typically not
be visible in a production environment.

This commit modifies Nova to log the message as a warning instead.

Change-Id: I347d2ab0077c4d9b62bc169af9e039d00ec214fa
Closes-bug: #1767553