Get anchors for sharing providers
Individual request groups (numbered or un-numbered) in a granular
resource request may be satisfied by a sharing provider.  In such cases,
we will need to be able to figure out which non-sharing trees the
sharing provider can be associated with.

This change introduces a helper method, _anchors_for_sharing_provider,
that accepts the ID of a sharing provider and returns the UUIDs of all
the unique root providers of trees associated with the same aggregate as
the sharing provider.

The sharing provider may or may not itself be part of a tree; in either
case, its root provider UUID is included in the result.

If the sharing provider is not part of any aggregate, the empty list is
returned.

blueprint: granular-resource-requests
Co-Authored-By: Jay Pipes <jaypipes@gmail.com>

Change-Id: I2cb4e95ca15caceb8713728f17630ed97059f25e