rework allocation handler _allocations_dict()
This is preparation for requiring consumer generation in API requests
and responses.

Reworks the utility functions for serialization in the allocation
handler in two ways: first, decouples the calculation of the last
modified timestamp from the serialization routines. second, breaks up
the _allocations_dict() megafunction into more readable (if slightly
more verbose) _serialize_allocations_for_consumer() and
_serialize_allocations_for_resource_provider() functions. It is these
functions whose docstrings actually describe the serialization output so
it's easier to understand what is going on without the conditionals for
last modified processing and the key_fetcher stuff.

Change-Id: I6f81b5af5400cf5e06a9d0b527699160da2a2ddc
blueprint: add-consumer-generation