Remove get_config_vhostuser
The 'get_config' in 'nova.virt.libvirt.vif' has two potential code
paths: the os-vif code path or the legacy code path. The former is tried
first, via a call to 'nova_to_osvif_vif' in 'nova.network.ovs_vif_util'.
This in turn looks for a function called '_nova_to_osvif_vif_XYZ', where
'XYZ' is the 'type' field from the VIF object. If this is found then the
relevant os-vif object is generated from nova's VIF objects and this is
used to call another function, '_set_config_XYZ', where 'XYZ' is the
type of the os-vif object. The legacy code path, by comparison, uses the
VIF object directly and simply attempts to call to 'get_config_XYZ' in
'nova.virt.libvirt.vif'. Assuming this is found, it is used to configure
both the VIF object and the libvirt XML.

For VIF types of 'vhostuser', the os-vif code path would attempt to
resolve '_nova_to_osvif_vif_vhostuser' in 'nova.network.os_vif_util'
followed by '_set_config_VIFVHostUser' in the same file. While both of
these exist, the former raises a 'NotImplementedError' for generic
vhost-user interfaces so the 'get_config_vhostuser' function is used,
for which we need to keep a lot of duplicated code around. Simplify this
by simply handling those generic vhost-user interfaces, allowing us to
resolve this issue.

Here's a mapping of the functions along with their os-vif calls:

  nova.virt.libvirt.vif._get_vhostuser_settings
    -> nova.network.os_vif_util._set_vhostuser_settings
  nova.virt.libvirt.vif.get_config_vhostuser
    -> nova.virt.libvirt.vif._set_config_VIFVHostUser

Change-Id: Ifab3006454708ab290b93f02d82b794c334c3946