pci: don't consider case when match tags specs
This makes matching pci device spec tags case
insensitive. This is useful to improve the user
experience for simple cases where boolean values
are being compared. For example, with the trusted
VF series, the port binding profile might have:

  trusted=True

And the PCI whitelist might have:

  trusted=true

And this should be considered a match.

Part of blueprint sriov-trusted-vfs

Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@redhat.com>
Change-Id: Ica532fa7dcd5fe11bd60db4cfe005a2bef0a1b7a