Expand member_of functional test cases
Now we support multiple member_of query in placement regardless of
whether sharing providers are in the environment or not.

However, the functional test is only for the cases where non-sharing
providers don't have the same resources with sharing providers.

This patch expands the existing test cases showing that we still
include allocation requests where the resource is provided by the
non-sharing providers when they have the same resources with sharing
providers.

Change-Id: Ide2272513c2f550e4539f2c94b581a82426f65f3
Blueprint: alloc-candidates-member-of