Fix being able to hard reboot a pausing instance
The allowed task states for a hard reboot include 'pausing'
but the instance.save(expected_task_states) doesn't include
'pausing', so if you try to hard reboot a pausing instance
it will fail with UnexpectedTaskStateError.

This makes the expected_task_states passed to Instance.save
use the same list of allowed task states that we use in the
check_instance_state decorator, and re-writes the unit test
to use an actual database to verify the Instance.save()
behavior and task state check in the DB API.

While we're at it, the API reference is updated to indicate
the allowed states for each type of reboot.

Change-Id: I5a21350e48a637e581d269fb567bb96c1899e174
Closes-Bug: #1768927