Add queued_for_delete field to InstanceMapping object
This patch adds a new field queued_for_delete to InstanceMapping
object which would be used to represent the value in the queued_for_delete
column of the instance_mappings table for each record.

This will be set to True if an instance is either deleted or soft_deleted.

Will be adding a data migration tool for filling this field.

Related to blueprint handling-down-cell

Change-Id: Iebb6974efda6dd5b24bc9ad987b6400455c780b6