Update queued-for-delete from the ComputeAPI during deletion/restoration
This patch updates queued-for-delete column in the instance_mapping
table from the ComputeAPI during normal delete, local delete (when
the compute service is down or instance is shelved offloaded) and
soft-delete operations. Note that as the name of the column suggests,
it is only "queued" for deletion and does not guarantee successful
deletion of the instance. Hence the value could be stale which is fine,
considering its use is only during down cell (desperate) situation.
This in no way hampers normal delete operations in nova.

This column is also updated from the ComputeAPI during the
nova restore operation.

Related to blueprint handling-down-cell

Change-Id: I3de5c839d212cfab1c69e2c78617c1051b971c38