Remove '_apply_instance_name_template'
Support for the 'multi_instance_display_name_template' option was
removed in commit 0e43002c9 and booting multiple instances now will
simply result in a simple numerical suffix (-1, -2, -3, ...) being added
to the instance's 'display_name' and 'hostname' attributes.

In that commit, it was noted that there was a lot more cleanup that
could be done in this area. The change gets that ball rolling by
squashing the '_apply_instance_name_template' function into
'_populate_instance_names'. This doesn't do much itself, but it allows
us to simplify a future change to remove much of the now-unnecessary
complexity of this feature.

Change-Id: I57ddfca39d9b76ac0cabf994d8a825a35b6d6e19